# improve import scanner perf