## Changes

Two big changes here to help the perf issues reported here: https://github.com/snowpackjs/snowpack/discussions/2896#discussioncomment-495424

1. Fixed our shockingly slow import replacer. This dropped `antd` first load down from ~700ms to ~300ms
2. Added memoization for our package import resolver. This dropped `antd` second load down from ~300ms to ~80ms. 
3. Bonus: Added `@ant-design/icons-svg` to the "never peer" list so that they get automatically bundled into any packages that request them. They're so small that this shouldn't seriously impact dev performance.

## Testing

- Covered by existing test suite.

## Docs

- N/A