@fgblomqvist 2021-03-18
Are those stats the time per import?