@FredKSchott 2021-03-18
That's for just the antd import, although that's a pretty special edge case. The bottleneck was on the "per import within the package" processing time, and antd has something like 50 different import statements to other sub-packages within it. It's a big package...