@Manish3323 2021-03-24
Hi @FredKSchott , it is still happening with our project for antd & ant-icons..
snowpack keeps on building package & never stops... 
here is the [repo](https://github.com/tmtsoftware/esw-ocs-eng-ui)
<img width="1123" alt="Screenshot 2021-03-24 at 11 32 13 PM" src="https://user-images.githubusercontent.com/16453816/112361296-3c384c80-8cf9-11eb-8c7d-df3003ac81b9.png">